// api/generated/models/calendarEventDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CalendarEventDto {
  id?: string;
  title?: string;
  startTime?: string;
  endTime?: string;
  description?: string;
  posterUrl?: string;
  ticketLink?: string;
  renterName?: string;
  engineers?: string[];
  contractUrl?: string;
}